Add files via upload
Created the codeBase_Group3.py file.
This is a python file that builds our best-performing language model. 
All unnecessary code has been commented out, after changing the file location of the training and test set the code can be run as it is, and it will print the perplexity of our language model (along with occasional updates on what the code is doing).